+++ reikna/cluda/cuda.py
@@ -12,7 +12,7 @@ from reikna.helpers import AttrDict
 import reikna.cluda as cluda
 import reikna.cluda.dtypes as dtypes
 from reikna.helpers import factors, wrap_in_tuple, product
-from reikna.cluda.kernel import render_prelude, render_template_source_with_modules
+from reikna.cluda.kernel import render_prelude, render_template_source
 from reikna.cluda.vsize import VirtualSizes, render_stub_vsize_funcs
 from reikna.cluda.tempalloc import ZeroOffsetManager
 
@@ -252,7 +252,7 @@ class Module:
         if render_kwds is None:
             render_kwds = {}
         prelude = render_prelude(self._ctx)
-        src = render_template_source_with_modules(src, **render_kwds)
+        src = render_template_source(src, **render_kwds)
 
         self.source = prelude + src
         self._module = ctx._compile(self.source)
@@ -335,7 +335,7 @@ class StaticKernel:
 
         prelude = render_prelude(self._ctx)
         stub_vsize_funcs = render_stub_vsize_funcs()
-        src = render_template_source_with_modules(template_src, *render_args, **render_kwds)
+        src = render_template_source(template_src, *render_args, **render_kwds)
 
         # We need the first approximation of the maximum thread number for a kernel.
         # Stub virtual size functions instead of real ones will not change it (hopefully).